Recreate missing Eclipse images based on context
`image03.png` and `image06.png` were recreated based on context in the
document.

For `image03.png` the text immediately prior to the image was used. It
claims that the image should show a toolbar, with the `Debug` button
highlighted in red.

`image06.png` had very little context was present. There are no direct
references to the image, meaning a guess had to be made.
Based on the previous sentences about debugging controls, it is assumed
that the image should show the debug bar with a red box around
"Step Into (F5)", "Step Over (F6)" and "Step Return (F7)".

Figure text was added to the images to make the reader aware that it is
not the original images.