Prevent known missing images from warning
The images spammed warnings such as
```
/src/en/Components/Network/wifi/wifi.rst:: WARNING: image file not readable: Components/Network/wifi/imgs/image1.png
```

Since the images are unavailable it is unlikely that they will ever not
be warnings.
Replacing the images with text specifying that an image is missing and
the reason for it both improves the experience for people reading the
documentation and reduces compiler warnings.

The missing images were previously just rendered as nothing, meaning
users reading the documentation would not know that an image was
missing.

The text has been bolded to stand out from other text on the site.