hardware_notes: Remove mention of 0x6xxx_xxxx and 0x7xxx_xxxx aliases
Not only was there a typo in this passage (saying writes were both
allowed and forbidden), it appears to have been entirely incorrect. Upon
retesting, I was not able to get either writes or reads to succeed for
ROM or RAM regions within these alleged aliases.

I probably made a mistake during my initial testing. Either that or the
older version of OpenOCD I was using then had a bug that has since been
fixed.